requirements.txt
adaptivecards==0.4.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docutils==0.21.2
executing==2.1.0
fastjsonschema==2.20.0
fqdn==1.5.1
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
json5==0.9.28
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==25.6.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.20
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.145
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.6.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.20
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.55.0
openapi-schema-pydantic==1.2.4
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
TAC-Insight @ file:///Users/vivekksi/PycharmProjects/TAC_Insight
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
twine==6.0.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
webex-bot==0.5.2
webexpythonsdk==2.0.3
webexteamssdk==1.6.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.13
yarl==1.18.0
zipp==3.21.0
